+++ project/src/components/map/map.tsx
@@ -1,7 +1,7 @@
 import 'leaflet/dist/leaflet.css';
 import { Offer } from '../../types/types';
 import { useEffect, useRef } from 'react';
-import  useMap from '../../hooks/useMap';
+import  useMap from '../../hooks/use-map/useMap';
 import {  LeafletUrls } from '../const';
 import { Marker } from 'leaflet';
 import { iconChanger } from '../../utils/utils';
